Hypervisor: add hypercall to ease loading of cache-colored images
Add a new hypercall that maps a colored memory region to the root cell so
that it can be seen as a contiguous region. This allows the usual load
function in the Jailhouse driver to be seamlessly used; it also avoid
duplicating the coloring logic by keeping it all in the hypervisor.

Signed-off-by: Luca Miccio <lucmiccio@gmail.com>
Signed-off-by: Marco Solieri <ms@xt3.it>