Configs: zynqmp-zcu102: add cache-coloring examples
Add colored root cell configuration for the ZCU102 evaluation board
along with two non root-cell configurations both for a bare metal inmate
and for a Linux one. In these configurations we are using the automated
colored allocation and we are assigning to each memory regions half of
the available colors on the platform.

Signed-off-by: Luca Miccio <lucmiccio@gmail.com>
Signed-off-by: Marco Solieri <ms@xt3.it>