Configs: jetson-tx2: add cache-coloring examples
Add colored root cell configuration for the NVIDIA TX2 development board
along with an example of a non root-cell configuration for a bare metal inmate.
In these configurations we are using the automated colored allocation and we
are assigning to each memory regions half of the available colors on the
platform.

Signed-off-by: Luca Miccio <lucmiccio@gmail.com>
Signed-off-by: Marco Solieri <ms@xt3.it>